% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bplfs-package.R
\docType{package}
\name{bplfs-package}
\alias{bplfs}
\alias{bplfs-package}
\title{bplfs: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
